+++ OAI/types/lora.py
@@ -1,8 +1,7 @@
 """ Lora types """
+from pydantic import BaseModel, Field
 from time import time
 from typing import Optional, List
-
-from pydantic import BaseModel, Field
 
 
 class LoraCard(BaseModel):
