+++ OAI/types/token.py
@@ -1,7 +1,6 @@
 """ Tokenization types """
-from typing import List
-
 from pydantic import BaseModel
+from typing import List
 
 
 class CommonTokenRequest(BaseModel):
