+++ common/templating.py
